requirements.txt
psycopg2-binary
Flask
python-dotenv
